feat: modify `selectCommand` to be utility function
