+++ .eslintrc.js
@@ -1,3 +1,7 @@
 const dev = require("@actions-kit/dev");
 
-module.exports = dev.createEslintConfig();
+module.exports = dev.createEslintConfig({
+  rules: {
+    "prettier/prettier": "off",
+  },
+});
