--- lib/commands/index.js
@@ -1,6 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.selectCommand = void 0;
-var root_1 = require("./root");
-Object.defineProperty(exports, "selectCommand", { enumerable: true, get: function () { return root_1.selectCommand; } });
-//# sourceMappingURL=index.js.map
