--- lib/commands/root.d.ts
@@ -1,1 +1,0 @@
-export declare function selectCommand(): Promise<void>;
