--- lib/commands/root.js
@@ -1,53 +1,0 @@
-"use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    var desc = Object.getOwnPropertyDescriptor(m, k);
-    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
-      desc = { enumerable: true, get: function() { return m[k]; } };
-    }
-    Object.defineProperty(o, k2, desc);
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || function (mod) {
-    if (mod && mod.__esModule) return mod;
-    var result = {};
-    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
-    __setModuleDefault(result, mod);
-    return result;
-};
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.selectCommand = void 0;
-const select_1 = __importDefault(require("@inquirer/select"));
-const venom = __importStar(require("venom-bot"));
-async function selectCommand() {
-    const trigger = await (0, select_1.default)({
-        message: "Select a command",
-        choices: [
-            {
-                name: "Run WhatsApp",
-                value: async () => {
-                    await venom.create({ headless: false, session: "default" });
-                },
-            },
-            {
-                name: "Exit",
-                value: () => {
-                    process.exit();
-                },
-            },
-        ],
-    });
-    await trigger();
-}
-exports.selectCommand = selectCommand;
-//# sourceMappingURL=root.js.map
