+++ lib/inputs.d.ts
@@ -1,0 +1,5 @@
+export interface Command {
+    name: string;
+    trigger: () => Promise<void>;
+}
+export declare function selectCommand(commands: Command[]): Promise<void>;
