+++ lib/whatscraper.js
@@ -24,9 +24,23 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-const commands = __importStar(require("./commands"));
+const venom = __importStar(require("venom-bot"));
+const inputs_1 = require("./inputs");
 async function launch() {
-    await commands.selectCommand();
+    await (0, inputs_1.selectCommand)([
+        {
+            name: "Run WhatsApp",
+            trigger: async () => {
+                await venom.create({ headless: false, session: "default" });
+            },
+        },
+        {
+            name: "Exit",
+            trigger: () => {
+                process.exit();
+            },
+        },
+    ]);
 }
 launch();
 //# sourceMappingURL=whatscraper.js.map
