+++ src/inputs.ts
@@ -1,0 +1,17 @@
+import select from "@inquirer/select";
+
+export interface Command {
+  name: string;
+  trigger: () => Promise<void>;
+}
+
+export async function selectCommand(commands: Command[]): Promise<void> {
+  const trigger = await select({
+    message: "Select a command",
+    choices: commands.map((command) => ({
+      name: command.name,
+      value: command.trigger,
+    })),
+  });
+  await trigger();
+}
