+++ src/whatscraper.ts
@@ -1,9 +1,23 @@
 #!/usr/bin/env node
 
-import * as commands from "./commands";
+import * as venom from "venom-bot";
+import { selectCommand } from "./inputs";
 
 async function launch() {
-  await commands.selectCommand();
+  await selectCommand([
+    {
+      name: "Run WhatsApp",
+      trigger: async () => {
+        await venom.create({ headless: false, session: "default" });
+      },
+    },
+    {
+      name: "Exit",
+      trigger: () => {
+        process.exit();
+      },
+    },
+  ]);
 }
 
 launch();
